+++ geo-traits/src/iterator.rs
@@ -4,13 +4,13 @@ use super::{
 };
 
 macro_rules! impl_iterator {
-    ($struct_name:ident, $self_trait:ident, $item_trait:ident, $access_method:ident, $item_type:ident) => {
+    ($struct_name:ident, $self_trait:ident, $item_trait:ident, $access_method:ident, $item_type:ident, $associated_type:ident) => {
         /// An iterator over the parts of this geometry.
         pub(crate) struct $struct_name<
             'a,
             T,
             $item_type: 'a + $item_trait<T = T>,
-            G: $self_trait<T = T, $item_type<'a> = $item_type>,
+            G: $self_trait<T = T, $associated_type<'a> = $item_type>,
         > {
             geom: &'a G,
             index: usize,
@@ -21,7 +21,7 @@ macro_rules! impl_iterator {
                 'a,
                 T,
                 $item_type: 'a + $item_trait<T = T>,
-                G: $self_trait<T = T, $item_type<'a> = $item_type>,
+                G: $self_trait<T = T, $associated_type<'a> = $item_type>,
             > $struct_name<'a, T, $item_type, G>
         {
             /// Create a new iterator
@@ -34,7 +34,7 @@ macro_rules! impl_iterator {
                 'a,
                 T,
                 $item_type: 'a + $item_trait<T = T>,
-                G: $self_trait<T = T, $item_type<'a> = $item_type>,
+                G: $self_trait<T = T, $associated_type<'a> = $item_type>,
             > Iterator for $struct_name<'a, T, $item_type, G>
         {
             type Item = $item_type;
@@ -59,7 +59,7 @@ macro_rules! impl_iterator {
                 'a,
                 T,
                 $item_type: 'a + $item_trait<T = T>,
-                G: $self_trait<T = T, $item_type<'a> = $item_type>,
+                G: $self_trait<T = T, $associated_type<'a> = $item_type>,
             > ExactSizeIterator for $struct_name<'a, T, $item_type, G>
         {
         }
@@ -68,7 +68,7 @@ macro_rules! impl_iterator {
                 'a,
                 T,
                 $item_type: 'a + $item_trait<T = T>,
-                G: $self_trait<T = T, $item_type<'a> = $item_type>,
+                G: $self_trait<T = T, $associated_type<'a> = $item_type>,
             > DoubleEndedIterator for $struct_name<'a, T, $item_type, G>
         {
             #[inline]
@@ -85,44 +85,50 @@ macro_rules! impl_iterator {
 }
 
 impl_iterator!(
-    LineStringIterator,
-    LineStringTrait,
-    CoordTrait,
-    coord_unchecked,
-    CoordType
+    LineStringIterator, // struct_name
+    LineStringTrait,    // self_trait
+    CoordTrait,         // item_trait
+    coord_unchecked,    // access_method
+    CoordType,          // item_type
+    CoordType           // associated_type
 );
 impl_iterator!(
-    PolygonInteriorIterator,
-    PolygonTrait,
-    LineStringTrait,
-    interior_unchecked,
-    RingType
+    PolygonInteriorIterator, // struct_name
+    PolygonTrait,            // self_trait
+    LineStringTrait,         // item_trait
+    interior_unchecked,      // access_method
+    RingType,                // item_type
+    RingType                 // associated_type
 );
 impl_iterator!(
-    MultiPointIterator,
-    MultiPointTrait,
-    PointTrait,
-    point_unchecked,
-    PointType
+    MultiPointIterator, // struct_name
+    MultiPointTrait,    // self_trait
+    PointTrait,         // item_trait
+    point_unchecked,    // access_method
+    PointType,          // item_type
+    InnerPointType      // associated_type
 );
 impl_iterator!(
-    MultiLineStringIterator,
-    MultiLineStringTrait,
-    LineStringTrait,
-    line_string_unchecked,
-    LineStringType
+    MultiLineStringIterator, // struct_name
+    MultiLineStringTrait,    // self_trait
+    LineStringTrait,         // item_trait
+    line_string_unchecked,   // access_method
+    LineStringType,          // item_type
+    InnerLineStringType      // associated_type
 );
 impl_iterator!(
-    MultiPolygonIterator,
-    MultiPolygonTrait,
-    PolygonTrait,
-    polygon_unchecked,
-    PolygonType
+    MultiPolygonIterator, // struct_name
+    MultiPolygonTrait,    // self_trait
+    PolygonTrait,         // item_trait
+    polygon_unchecked,    // access_method
+    PolygonType,          // item_type
+    InnerPolygonType      // associated_type
 );
 impl_iterator!(
-    GeometryCollectionIterator,
-    GeometryCollectionTrait,
-    GeometryTrait,
-    geometry_unchecked,
-    GeometryType
+    GeometryCollectionIterator, // struct_name
+    GeometryCollectionTrait,    // self_trait
+    GeometryTrait,              // item_trait
+    geometry_unchecked,         // access_method
+    GeometryType,               // item_type
+    GeometryType                // associated_type
 );