+++ geo-traits/src/multi_line_string.rs
@@ -2,7 +2,7 @@ use std::marker::PhantomData;
 
 use crate::iterator::MultiLineStringIterator;
 use crate::line_string::UnimplementedLineString;
-use crate::{Dimensions, LineStringTrait};
+use crate::{GeometryTrait, LineStringTrait};
 #[cfg(feature = "geo-types")]
 use geo_types::{CoordNum, LineString, MultiLineString};
 
@@ -11,22 +11,16 @@ use geo_types::{CoordNum, LineString, MultiLineString};
 /// A MultiLineString is a collection of [`LineString`s][LineStringTrait].
 ///
 /// Refer to [geo_types::MultiLineString] for information about semantics and validity.
-pub trait MultiLineStringTrait: Sized {
-    /// The coordinate type of this geometry
-    type T;
-
+pub trait MultiLineStringTrait: Sized + GeometryTrait {
     /// The type of each underlying LineString, which implements [LineStringTrait]
-    type LineStringType<'a>: 'a + LineStringTrait<T = Self::T>
+    type InnerLineStringType<'a>: 'a + LineStringTrait<T = Self::T>
     where
         Self: 'a;
-
-    /// The dimension of this geometry
-    fn dim(&self) -> Dimensions;
 
     /// An iterator over the LineStrings in this MultiLineString
     fn line_strings(
         &self,
-    ) -> impl DoubleEndedIterator + ExactSizeIterator<Item = Self::LineStringType<'_>> {
+    ) -> impl DoubleEndedIterator + ExactSizeIterator<Item = Self::InnerLineStringType<'_>> {
         MultiLineStringIterator::new(self, 0, self.num_line_strings())
     }
 
@@ -35,7 +29,7 @@ pub trait MultiLineStringTrait: Sized {
 
     /// Access to a specified line_string in this MultiLineString
     /// Will return None if the provided index is out of bounds
-    fn line_string(&self, i: usize) -> Option<Self::LineStringType<'_>> {
+    fn line_string(&self, i: usize) -> Option<Self::InnerLineStringType<'_>> {
         if i >= self.num_line_strings() {
             None
         } else {
@@ -48,47 +42,37 @@ pub trait MultiLineStringTrait: Sized {
     /// # Safety
     ///
     /// Accessing an index out of bounds is UB.
-    unsafe fn line_string_unchecked(&self, i: usize) -> Self::LineStringType<'_>;
+    unsafe fn line_string_unchecked(&self, i: usize) -> Self::InnerLineStringType<'_>;
 }
 
 #[cfg(feature = "geo-types")]
 impl<T: CoordNum> MultiLineStringTrait for MultiLineString<T> {
-    type T = T;
-    type LineStringType<'a>
+    type InnerLineStringType<'a>
         = &'a LineString<Self::T>
     where
         Self: 'a;
-
-    fn dim(&self) -> Dimensions {
-        Dimensions::Xy
-    }
 
     fn num_line_strings(&self) -> usize {
         self.0.len()
     }
 
-    unsafe fn line_string_unchecked(&self, i: usize) -> Self::LineStringType<'_> {
+    unsafe fn line_string_unchecked(&self, i: usize) -> Self::InnerLineStringType<'_> {
         self.0.get_unchecked(i)
     }
 }
 
 #[cfg(feature = "geo-types")]
 impl<'a, T: CoordNum> MultiLineStringTrait for &'a MultiLineString<T> {
-    type T = T;
-    type LineStringType<'b>
+    type InnerLineStringType<'b>
         = &'a LineString<Self::T>
     where
         Self: 'b;
-
-    fn dim(&self) -> Dimensions {
-        Dimensions::Xy
-    }
 
     fn num_line_strings(&self) -> usize {
         self.0.len()
     }
 
-    unsafe fn line_string_unchecked(&self, i: usize) -> Self::LineStringType<'_> {
+    unsafe fn line_string_unchecked(&self, i: usize) -> Self::InnerLineStringType<'_> {
         self.0.get_unchecked(i)
     }
 }
@@ -100,21 +84,16 @@ impl<'a, T: CoordNum> MultiLineStringTrait for &'a MultiLineString<T> {
 pub struct UnimplementedMultiLineString<T>(PhantomData<T>);
 
 impl<T> MultiLineStringTrait for UnimplementedMultiLineString<T> {
-    type T = T;
-    type LineStringType<'a>
+    type InnerLineStringType<'a>
         = UnimplementedLineString<Self::T>
     where
         Self: 'a;
-
-    fn dim(&self) -> Dimensions {
-        unimplemented!()
-    }
 
     fn num_line_strings(&self) -> usize {
         unimplemented!()
     }
 
-    unsafe fn line_string_unchecked(&self, _i: usize) -> Self::LineStringType<'_> {
+    unsafe fn line_string_unchecked(&self, _i: usize) -> Self::InnerLineStringType<'_> {
         unimplemented!()
     }
 }