+++ geo-traits/src/multi_point.rs
@@ -1,7 +1,7 @@
 use std::marker::PhantomData;
 
 use crate::iterator::MultiPointIterator;
-use crate::{Dimensions, PointTrait, UnimplementedPoint};
+use crate::{GeometryTrait, PointTrait, UnimplementedPoint};
 #[cfg(feature = "geo-types")]
 use geo_types::{CoordNum, MultiPoint, Point};
 
@@ -10,20 +10,16 @@ use geo_types::{CoordNum, MultiPoint, Point};
 /// A MultiPoint is a collection of [`Point`s][PointTrait].
 ///
 /// Refer to [geo_types::MultiPoint] for information about semantics and validity.
-pub trait MultiPointTrait: Sized {
-    /// The coordinate type of this geometry
-    type T;
-
+pub trait MultiPointTrait: Sized + GeometryTrait {
     /// The type of each underlying Point, which implements [PointTrait]
-    type PointType<'a>: 'a + PointTrait<T = Self::T>
+    type InnerPointType<'a>: 'a + PointTrait<T = Self::T>
     where
         Self: 'a;
 
-    /// The dimension of this geometry
-    fn dim(&self) -> Dimensions;
-
     /// An iterator over the points in this MultiPoint
-    fn points(&self) -> impl DoubleEndedIterator + ExactSizeIterator<Item = Self::PointType<'_>> {
+    fn points(
+        &self,
+    ) -> impl DoubleEndedIterator + ExactSizeIterator<Item = Self::InnerPointType<'_>> {
         MultiPointIterator::new(self, 0, self.num_points())
     }
 
@@ -32,7 +28,7 @@ pub trait MultiPointTrait: Sized {
 
     /// Access to a specified point in this MultiPoint
     /// Will return None if the provided index is out of bounds
-    fn point(&self, i: usize) -> Option<Self::PointType<'_>> {
+    fn point(&self, i: usize) -> Option<Self::InnerPointType<'_>> {
         if i >= self.num_points() {
             None
         } else {
@@ -45,47 +41,37 @@ pub trait MultiPointTrait: Sized {
     /// # Safety
     ///
     /// Accessing an index out of bounds is UB.
-    unsafe fn point_unchecked(&self, i: usize) -> Self::PointType<'_>;
+    unsafe fn point_unchecked(&self, i: usize) -> Self::InnerPointType<'_>;
 }
 
 #[cfg(feature = "geo-types")]
 impl<T: CoordNum> MultiPointTrait for MultiPoint<T> {
-    type T = T;
-    type PointType<'a>
+    type InnerPointType<'a>
         = &'a Point<Self::T>
     where
         Self: 'a;
-
-    fn dim(&self) -> Dimensions {
-        Dimensions::Xy
-    }
 
     fn num_points(&self) -> usize {
         self.0.len()
     }
 
-    unsafe fn point_unchecked(&self, i: usize) -> Self::PointType<'_> {
+    unsafe fn point_unchecked(&self, i: usize) -> Self::InnerPointType<'_> {
         self.0.get_unchecked(i)
     }
 }
 
 #[cfg(feature = "geo-types")]
 impl<'a, T: CoordNum> MultiPointTrait for &'a MultiPoint<T> {
-    type T = T;
-    type PointType<'b>
+    type InnerPointType<'b>
         = &'a Point<Self::T>
     where
         Self: 'b;
-
-    fn dim(&self) -> Dimensions {
-        Dimensions::Xy
-    }
 
     fn num_points(&self) -> usize {
         self.0.len()
     }
 
-    unsafe fn point_unchecked(&self, i: usize) -> Self::PointType<'_> {
+    unsafe fn point_unchecked(&self, i: usize) -> Self::InnerPointType<'_> {
         self.0.get_unchecked(i)
     }
 }
@@ -97,21 +83,16 @@ impl<'a, T: CoordNum> MultiPointTrait for &'a MultiPoint<T> {
 pub struct UnimplementedMultiPoint<T>(PhantomData<T>);
 
 impl<T> MultiPointTrait for UnimplementedMultiPoint<T> {
-    type T = T;
-    type PointType<'a>
+    type InnerPointType<'a>
         = UnimplementedPoint<Self::T>
     where
         Self: 'a;
-
-    fn dim(&self) -> Dimensions {
-        unimplemented!()
-    }
 
     fn num_points(&self) -> usize {
         unimplemented!()
     }
 
-    unsafe fn point_unchecked(&self, _i: usize) -> Self::PointType<'_> {
+    unsafe fn point_unchecked(&self, _i: usize) -> Self::InnerPointType<'_> {
         unimplemented!()
     }
 }