+++ geo-traits/src/multi_polygon.rs
@@ -2,29 +2,23 @@ use std::marker::PhantomData;
 
 use crate::iterator::MultiPolygonIterator;
 use crate::polygon::UnimplementedPolygon;
-use crate::{Dimensions, PolygonTrait};
+use crate::{GeometryTrait, PolygonTrait};
 #[cfg(feature = "geo-types")]
 use geo_types::{CoordNum, MultiPolygon, Polygon};
 
 /// A trait for accessing data from a generic MultiPolygon.
 ///
 /// Refer to [geo_types::MultiPolygon] for information about semantics and validity.
-pub trait MultiPolygonTrait: Sized {
-    /// The coordinate type of this geometry
-    type T;
-
+pub trait MultiPolygonTrait: Sized + GeometryTrait {
     /// The type of each underlying Polygon, which implements [PolygonTrait]
-    type PolygonType<'a>: 'a + PolygonTrait<T = Self::T>
+    type InnerPolygonType<'a>: 'a + PolygonTrait<T = Self::T>
     where
         Self: 'a;
-
-    /// The dimension of this geometry
-    fn dim(&self) -> Dimensions;
 
     /// An iterator over the Polygons in this MultiPolygon
     fn polygons(
         &self,
-    ) -> impl DoubleEndedIterator + ExactSizeIterator<Item = Self::PolygonType<'_>> {
+    ) -> impl DoubleEndedIterator + ExactSizeIterator<Item = Self::InnerPolygonType<'_>> {
         MultiPolygonIterator::new(self, 0, self.num_polygons())
     }
 
@@ -33,7 +27,7 @@ pub trait MultiPolygonTrait: Sized {
 
     /// Access to a specified polygon in this MultiPolygon
     /// Will return None if the provided index is out of bounds
-    fn polygon(&self, i: usize) -> Option<Self::PolygonType<'_>> {
+    fn polygon(&self, i: usize) -> Option<Self::InnerPolygonType<'_>> {
         if i >= self.num_polygons() {
             None
         } else {
@@ -46,47 +40,37 @@ pub trait MultiPolygonTrait: Sized {
     /// # Safety
     ///
     /// Accessing an index out of bounds is UB.
-    unsafe fn polygon_unchecked(&self, i: usize) -> Self::PolygonType<'_>;
+    unsafe fn polygon_unchecked(&self, i: usize) -> Self::InnerPolygonType<'_>;
 }
 
 #[cfg(feature = "geo-types")]
 impl<T: CoordNum> MultiPolygonTrait for MultiPolygon<T> {
-    type T = T;
-    type PolygonType<'a>
+    type InnerPolygonType<'a>
         = &'a Polygon<Self::T>
     where
         Self: 'a;
-
-    fn dim(&self) -> Dimensions {
-        Dimensions::Xy
-    }
 
     fn num_polygons(&self) -> usize {
         self.0.len()
     }
 
-    unsafe fn polygon_unchecked(&self, i: usize) -> Self::PolygonType<'_> {
+    unsafe fn polygon_unchecked(&self, i: usize) -> Self::InnerPolygonType<'_> {
         self.0.get_unchecked(i)
     }
 }
 
 #[cfg(feature = "geo-types")]
 impl<'a, T: CoordNum> MultiPolygonTrait for &'a MultiPolygon<T> {
-    type T = T;
-    type PolygonType<'b>
+    type InnerPolygonType<'b>
         = &'a Polygon<Self::T>
     where
         Self: 'b;
-
-    fn dim(&self) -> Dimensions {
-        Dimensions::Xy
-    }
 
     fn num_polygons(&self) -> usize {
         self.0.len()
     }
 
-    unsafe fn polygon_unchecked(&self, i: usize) -> Self::PolygonType<'_> {
+    unsafe fn polygon_unchecked(&self, i: usize) -> Self::InnerPolygonType<'_> {
         self.0.get_unchecked(i)
     }
 }
@@ -98,21 +82,16 @@ impl<'a, T: CoordNum> MultiPolygonTrait for &'a MultiPolygon<T> {
 pub struct UnimplementedMultiPolygon<T>(PhantomData<T>);
 
 impl<T> MultiPolygonTrait for UnimplementedMultiPolygon<T> {
-    type T = T;
-    type PolygonType<'a>
+    type InnerPolygonType<'a>
         = UnimplementedPolygon<Self::T>
     where
         Self: 'a;
-
-    fn dim(&self) -> Dimensions {
-        unimplemented!()
-    }
 
     fn num_polygons(&self) -> usize {
         unimplemented!()
     }
 
-    unsafe fn polygon_unchecked(&self, _i: usize) -> Self::PolygonType<'_> {
+    unsafe fn polygon_unchecked(&self, _i: usize) -> Self::InnerPolygonType<'_> {
         unimplemented!()
     }
 }